Move extended TCP keepalive parameters to TCPSSLOptions, apply in EpollTransport (only)
